refactored setting the status bar metrics
