+++ src/com/swdc/netbeans/plugin/status/SoftwareStatusBar.java
@@ -32,21 +32,16 @@ public class SoftwareStatusBar implements StatusLineElementProvider {
     private long last_click_time = -1;
     
     private static boolean showStatusText = true;
-    private static StatusBarType lastStatusType = StatusBarType.NO_KPM;
+    private static StatusBarType lastStatusType = StatusBarType.PAW;
     private static String lastMsg = "";
     private static String lastTooltip = "";
 
     private static boolean registeredMouseClick = false;
     
-    public enum StatusBarType {
-        ROCKET("com/swdc/netbeans/plugin/status/rocket.png"),
-        FULL("com/swdc/netbeans/plugin/status/100.png"),
-        ALMOST("com/swdc/netbeans/plugin/status/75.png"),
-        HALF("com/swdc/netbeans/plugin/status/50.png"),
-        QUARTER("com/swdc/netbeans/plugin/status/25.png"),
-        NO_KPM("com/swdc/netbeans/plugin/status/sw.png"),
-        ALERT("com/swdc/netbeans/plugin/status/warning.png"),
-        OFF("com/swdc/netbeans/plugin/status/clock.png");
+    public static enum StatusBarType {
+        PAW("com/swdc/netbeans/plugin/assets/paw-grey.png"),
+        ROCKET("com/swdc/netbeans/plugin/assets/rocket.png"),
+        OFF("com/swdc/netbeans/plugin/assets/clock.png");
 
         private Icon icon;
 
@@ -57,7 +52,7 @@ public class SoftwareStatusBar implements StatusLineElementProvider {
 
     public SoftwareStatusBar() {
         
-        statusLabel.setIcon(StatusBarType.NO_KPM.icon);
+        statusLabel.setIcon(StatusBarType.PAW.icon);
         if (!registeredMouseClick) {
             registeredMouseClick = true;
             statusLabel.addMouseListener(new MouseAdapter() {
@@ -77,7 +72,7 @@ public class SoftwareStatusBar implements StatusLineElementProvider {
             updateMessage(lastStatusType, lastMsg, lastTooltip);
         } else {
             // set status line will update it to a clock
-            updateMessage(StatusBarType.NO_KPM, "", lastTooltip);
+            updateMessage(StatusBarType.OFF, "", lastTooltip);
         }
     }
 
@@ -107,26 +102,11 @@ public class SoftwareStatusBar implements StatusLineElementProvider {
             case ROCKET:
                 statusLabel.setIcon(StatusBarType.ROCKET.icon);
                 break;
-            case FULL:
-                statusLabel.setIcon(StatusBarType.FULL.icon);
-                break;
-            case ALMOST:
-                statusLabel.setIcon(StatusBarType.ALMOST.icon);
-                break;
-            case HALF:
-                statusLabel.setIcon(StatusBarType.HALF.icon);
-                break;
-            case QUARTER:
-                statusLabel.setIcon(StatusBarType.QUARTER.icon);
-                break;
-            case ALERT:
-                statusLabel.setIcon(StatusBarType.ALERT.icon);
-                break;
             case OFF:
                 statusLabel.setIcon(StatusBarType.OFF.icon);
                 break;
             default:
-                statusLabel.setIcon(StatusBarType.NO_KPM.icon);
+                statusLabel.setIcon(StatusBarType.PAW.icon);
                 break;
         }
         statusLabel.repaint();
